Moved `get_lowqual_expr` from variant_qc.py to annotations.py
